Support frontmatter to exclude page from sitemap
This PR relates to #18. I've implemented support of parsing frontmatter data to exclude a page from sitemap. 

Like in the original feature request, you can set `sitemap: false`to exclude a specific page. It further checks for the meta data and searches for a `robots` value containing `noindex`.

Currently, if above data evaluates to true, the page path will be pushed to the exclude option parameter. It could be discussed, wether to break / return from the forEach loop instead.